requirements.txt
yfinance
pandas==1.3.5
plotly_express==0.4.0
plotly==5.7.0
matplotlib==3.5.1
numpy==1.21.6
streamlit-lottie==0.0.3
cryptocompare
streamlit-option-menu